requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.2.0
constantly==15.1.0
cryptography==41.0.2
daphne==4.0.0
Django==4.1.5
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
Pillow==10.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
stripe==5.0.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
urllib3==2.0.3
whitenoise==6.4.0
zope.interface==6.0
